changed file names, and removed getUser from auth to user route
