+++ src/api/auth/auth.controller.js
@@ -1,17 +1,7 @@
 const bcrypt = require('bcryptjs');
 
-const User = require('../user/model.js');
-const authHelper = require('./helper.js');
-
-const authorizeUser = async (req, res) => {
-  try {
-    const user = await User.findById(req.user.id).select('-password');
-    res.status(200).json({ user });
-  } catch (err) {
-    console.log(err.message);
-    res.status(500).send('Server Error');
-  }
-};
+const User = require('../user/user.model');
+const authHelper = require('./auth.helper');
 
 const authenticateUser = async (req, res) => {
   try {
@@ -27,14 +17,13 @@ const authenticateUser = async (req, res) => {
     }
 
     const token = authHelper.createJwtToken(user);
-    res.status(200).json({ token });
+    return res.status(200).json({ token });
   } catch (err) {
     console.log(err.message);
-    res.status(500).send('Server Error');
+    return res.status(500).send('Server Error');
   }
 };
 
 module.exports = {
-  authorizeUser,
   authenticateUser,
 };
